chore: adds new step with address details
