+++ libs/dh/globalization/assets-localization/src/assets/i18n/en.json
@@ -674,8 +674,10 @@
           "attention": "Attn.",
           "phoneNumber": "Phone number",
           "mobile": "Mobile",
-          "email": "Email address",
-          "addressDetails": "Address details",
+          "email": "Email address"
+        },
+        "addressDetails": {
+          "label": "Address details",
           "addressSameAsMeteringPoint": "Same as metering point address",
           "street": "Street",
           "houseNumber": "House number",
