+++ libs/dh/metering-point/feature-move-in/src/components/move-in.component.html
@@ -33,9 +33,20 @@ limitations under the License.
       [stepControl]="contactDetailsForm"
       [label]="t('steps.contactDetails.label')"
       [previousButtonLabel]="t('steps.customerDetails.label')"
-      [nextButtonLabel]="t('save')"
+      [nextButtonLabel]="t('steps.addressDetails.label')"
     >
       <dh-contact-details-form [contactDetailsForm]="contactDetailsForm" />
+    </watt-stepper-step>
+
+    <watt-stepper-step
+      [stepControl]="contactDetailsForm"
+      [label]="t('steps.addressDetails.label')"
+      [previousButtonLabel]="t('steps.contactDetails.label')"
+      [nextButtonLabel]="t('save')"
+    >
+      <dh-address-details-form [addressDetailsForm]="addressDetailsForm"
+                               (resetLegalForm)="emptyLegalAddressFormGroup()"
+                               (resetTechnicalForm)="emptyTechnicalAddressFormGroup()" />
     </watt-stepper-step>
   </watt-stepper>
 </watt-modal>
